made open web view a reusable function
